refactor(mark): move usability test client
I am going to try putting a usability test app in here too, and want them to live together.
